Move official pool mappings to separate script, and add official pool flag instead od new pool type
