+++ manifest.template.yaml
@@ -2122,7 +2122,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.5
       language: wasm/assemblyscript
-      file: ./src/mappings/poolController.ts
+      file: ./src/mappings/poolOfficial.ts
       entities:
         - PoolContract
       abis:
