+++ schema.graphql
@@ -27,6 +27,7 @@ type Pool @entity {
   swapFee: BigDecimal!
   owner: Bytes
   isPaused: Boolean
+  officialPool: Boolean!
 
   totalWeight: BigDecimal
   totalSwapVolume: BigDecimal!
