+++ src/mappings/poolFactory.ts
@@ -62,8 +62,6 @@ import { GyroEPool } from '../types/templates/GyroEPool/GyroEPool';
 import { Transfer } from '../types/Vault/ERC20';
 import { handleTransfer } from './poolController';
 import { ComposableStablePool } from '../types/ComposableStablePoolFactory/ComposableStablePool';
-import { OfficialPoolRegistered } from '../types/templates/OfficialPoolsRegister/OfficialPoolsRegister';
-import { OfficialPoolDeregistered } from '../types/templates/OfficialPoolsRegister/OfficialPoolsRegister';
 
 function createWeightedLikePool(event: PoolCreated, poolType: string, poolTypeVersion: i32 = 1): string | null {
   let poolAddress: Address = event.params.pool;
@@ -653,6 +651,7 @@ function handleNewPool(event: PoolCreated, poolId: Bytes, swapFee: BigInt): Pool
     pool.tx = event.transaction.hash;
     pool.swapEnabled = true;
     pool.isPaused = false;
+    pool.officialPool = false;
 
     let bpt = getToken(poolAddress);
 
@@ -703,20 +702,4 @@ export function handleProtocolIdRegistryOrRename(event: ProtocolIdRegistered): v
     protocol.name = event.params.name;
   }
   protocol.save();
-}
-
-export function handleOfficialPoolRegistered(event: OfficialPoolRegistered): void {
-  let poolId = event.params.poolId;
-  let pool = Pool.load(poolId.toHexString()) as Pool;
-  pool.poolType = PoolType.OfficialPool;
-
-  pool.save();
-}
-
-export function handleOfficalPoolDeregistered(event: OfficialPoolDeregistered): void {
-  let poolId = event.params.poolId;
-  let pool = Pool.load(poolId.toHexString()) as Pool;
-  pool.poolType = PoolType.OfficialPool;
-
-  pool.save();
 }
